# atoma-proxy/docs/openapi.yml
openapi: 3.1.0
info:
  title: atoma-proxy
  description: ''
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
servers:
- url: http://localhost:8080
paths:
...
  /v1/images/generations:
    post:
      tags:
      - Images
      summary: Create image generation
      description: |-
        This endpoint processes requests to generate images using AI models by forwarding them
        to the appropriate AI node. The request metadata and compute units have already been
        validated by middleware before reaching this handler.

        # Arguments
        * `metadata` - Extension containing pre-processed request metadata (node address, compute units, etc.)
        * `state` - Application state containing configuration and shared resources
        * `headers` - HTTP headers from the incoming request
        * `payload` - JSON payload containing image generation parameters

        # Returns
        * `Result<Response<Body>>` - The processed response from the AI node or an error status

        # Errors
        * Returns various status codes based on the underlying `handle_image_generation_response`:
          - `INTERNAL_SERVER_ERROR` - If there's an error communicating with the AI node

        # Example Payload
        ```json
        {
            "model": "stable-diffusion-v1-5",
            "n": 1,
            "size": "1024x1024"
        }
        ```
      operationId: image_generations_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageRequest'
        required: true
      responses:
        '200':
          description: Image generations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateImageResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
  /v1/confidential/images/generations:
    post:
      tags:
      - Confidential Images
      summary: Create confidential image generations
      description: |-
        This endpoint follows the OpenAI API format for generating images,
        but with confidential processing (through AEAD encryption and TEE hardware).
        The handler receives pre-processed metadata from middleware and forwards the request to
        the selected node.

        Note: Authentication, node selection, initial request validation and encryption
        are handled by middleware before this handler is called.

        # Arguments
        * `metadata` - Pre-processed request metadata containing node information and compute units
        * `state` - The shared proxy state containing configuration and runtime information
        * `headers` - HTTP headers from the incoming request
        * `payload` - The JSON request body containing the model and input text

        # Returns
        * `Ok(Response)` - The image generations response from the processing node
        * `Err(AtomaProxyError)` - An error status code if any step fails

        # Errors
        * `INTERNAL_SERVER_ERROR` - Processing or node communication failures
      operationId: confidential_image_generations_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageRequest'
        required: true
      responses:
        '200':
          description: Image generations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateImageResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    CreateImageRequest:
      type: object
      description: Request body for image generation
      required:
      - prompt
      - model
      - n
      properties:
        model:
          type: string
          description: The model to use for image generation.
        n:
          type: integer
          format: int32
          description: The number of images to generate. Must be between 1 and 10.
          minimum: 0
        prompt:
          type: string
          description: A text description of the desired image(s). The maximum length is 1000 characters.
        quality:
          type:
          - string
          - 'null'
          description: |-
            The quality of the image that will be generated.
            `hd` creates images with finer details and greater consistency across the image.
        response_format:
          type:
          - string
          - 'null'
          description: The format in which the generated images are returned.
        size:
          type:
          - string
          - 'null'
          description: The size of the generated images.
        style:
          type:
          - string
          - 'null'
          description: The style of the generated images.
        user:
          type:
          - string
          - 'null'
          description: A unique identifier representing your end-user, which can help OpenAI to monitor and detect abuse.
    CreateImageResponse:
      type: object
      description: Response format for image generation
      required:
      - created
      - data
      properties:
        created:
          type: integer
          format: int64
        data:
          type: array
          items:
            $ref: '#/components/schemas/ImageData'
...
    ImageData:
      type: object
      description: Individual image data in the response
      required:
      - url
      properties:
        revised_prompt:
          type:
          - string
          - 'null'
        url:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...